Move data layer files into correct folders and packages
