+++ app/src/main/java/com/example/android/architecture/blueprints/todoapp/data/DefaultTasksRepository.kt
@@ -14,14 +14,10 @@
  * limitations under the License.
  */
 
-package com.example.android.architecture.blueprints.todoapp.data.source
+package com.example.android.architecture.blueprints.todoapp.data
 
-import com.example.android.architecture.blueprints.todoapp.data.Task
+import com.example.android.architecture.blueprints.todoapp.data.source.network.NetworkDataSource
 import com.example.android.architecture.blueprints.todoapp.data.source.local.TasksDao
-import com.example.android.architecture.blueprints.todoapp.data.toExternalModel
-import com.example.android.architecture.blueprints.todoapp.data.toLocalModel
-import com.example.android.architecture.blueprints.todoapp.data.toNetworkModel
-import com.example.android.architecture.blueprints.todoapp.data.toTaskEntity
 import kotlinx.coroutines.coroutineScope
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
