--- app/src/main/java/com/example/android/architecture/blueprints/todoapp/data/source/NetworkDataSource.kt
@@ -1,42 +1,0 @@
-/*
- * Copyright 2019 The Android Open Source Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.example.android.architecture.blueprints.todoapp.data.source
-
-import com.example.android.architecture.blueprints.todoapp.data.source.remote.NetworkTask
-
-/**
- * Main entry point for accessing tasks data from the network.
- *
- */
-interface NetworkDataSource {
-
-    suspend fun loadTasks(): List<NetworkTask>
-
-    suspend fun getTask(taskId: String): NetworkTask?
-
-    suspend fun saveTask(task: NetworkTask)
-
-    suspend fun completeTask(taskId: String)
-
-    suspend fun activateTask(taskId: String)
-
-    suspend fun clearCompletedTasks()
-
-    suspend fun deleteAllTasks()
-
-    suspend fun deleteTask(taskId: String)
-}
