+++ app/src/main/java/com/example/android/architecture/blueprints/todoapp/data/source/network/NetworkDataSource.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2023 The Android Open Source Project
+ * Copyright 2019 The Android Open Source Project
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,24 +14,27 @@
  * limitations under the License.
  */
 
-package com.example.android.architecture.blueprints.todoapp.data.source.remote
+package com.example.android.architecture.blueprints.todoapp.data.source.network
 
 /**
- * Internal model used to represent a task obtained from the network. This is used inside the data
- * layer only.
+ * Main entry point for accessing tasks data from the network.
  *
- * See ModelMappingExt.kt for mapping functions used to convert this model to other
- * models.
  */
-data class NetworkTask(
-    val id: String,
-    val title: String,
-    val shortDescription: String,
-    val priority: Int? = null,
-    val status: TaskStatus = TaskStatus.ACTIVE
-)
+interface NetworkDataSource {
 
-enum class TaskStatus {
-    ACTIVE,
-    COMPLETE
+    suspend fun loadTasks(): List<NetworkTask>
+
+    suspend fun getTask(taskId: String): NetworkTask?
+
+    suspend fun saveTask(task: NetworkTask)
+
+    suspend fun completeTask(taskId: String)
+
+    suspend fun activateTask(taskId: String)
+
+    suspend fun clearCompletedTasks()
+
+    suspend fun deleteAllTasks()
+
+    suspend fun deleteTask(taskId: String)
 }
